A few further improvements to the HDF5 writing capability. Split the (very long) convert detail.hpp function into several parts.
